+++ tests/unit/test_models.py
@@ -11,11 +11,19 @@ class TestModels:
     A class to test the models.
     """
 
-    @pytest.mark.unit
-    def test_simulate_without_build_model(self):
-        # Define model
-        model = pybop.lithium_ion.SPM()
+    @pytest.fixture(
+        params=[
+            pybop.lithium_ion.SPM(),
+            pybop.lithium_ion.SPMe(),
+            pybop.empirical.Thevenin(),
+        ]
+    )
+    def model(self, request):
+        model = request.param
+        return model.copy()
 
+    @pytest.mark.unit
+    def test_simulate_without_build_model(self, model):
         with pytest.raises(
             ValueError, match="Model must be built before calling simulate"
         ):
@@ -27,49 +35,47 @@ class TestModels:
             model.simulateS1(None, None)
 
     @pytest.mark.unit
-    def test_predict_without_pybamm(self):
-        # Define model
-        model = pybop.lithium_ion.SPM()
+    def test_predict_without_pybamm(self, model):
         model._unprocessed_model = None
 
         with pytest.raises(ValueError):
             model.predict(None, None)
 
     @pytest.mark.unit
-    def test_predict_with_inputs(self):
-        # Define SPM
-        model = pybop.lithium_ion.SPM()
+    def test_predict_with_inputs(self, model):
+        # Define inputs
         t_eval = np.linspace(0, 10, 100)
-        inputs = {
-            "Negative electrode active material volume fraction": 0.52,
-            "Positive electrode active material volume fraction": 0.63,
-        }
+        if isinstance(model, (pybop.lithium_ion.SPM, pybop.lithium_ion.SPMe)):
+            inputs = {
+                "Negative electrode active material volume fraction": 0.52,
+                "Positive electrode active material volume fraction": 0.63,
+            }
+        elif isinstance(model, (pybop.empirical.Thevenin)):
+            inputs = {
+                "R0 [Ohm]": 0.0002,
+                "R1 [Ohm]": 0.0001,
+            }
+        else:
+            raise ValueError("Inputs not defined for this type of model.")
 
         res = model.predict(t_eval=t_eval, inputs=inputs)
-        assert len(res["Terminal voltage [V]"].data) == 100
-
-        # Define SPMe
-        model = pybop.lithium_ion.SPMe()
-        res = model.predict(t_eval=t_eval, inputs=inputs)
-        assert len(res["Terminal voltage [V]"].data) == 100
+        assert len(res["Voltage [V]"].data) == 100
 
     @pytest.mark.unit
-    def test_predict_without_allow_infeasible_solutions(self):
-        # Define SPM
-        model = pybop.lithium_ion.SPM()
-        model.allow_infeasible_solutions = False
-        t_eval = np.linspace(0, 10, 100)
-        inputs = {
-            "Negative electrode active material volume fraction": 0.9,
-            "Positive electrode active material volume fraction": 0.9,
-        }
+    def test_predict_without_allow_infeasible_solutions(self, model):
+        if isinstance(model, (pybop.lithium_ion.SPM, pybop.lithium_ion.SPMe)):
+            model.allow_infeasible_solutions = False
+            t_eval = np.linspace(0, 10, 100)
+            inputs = {
+                "Negative electrode active material volume fraction": 0.9,
+                "Positive electrode active material volume fraction": 0.9,
+            }
 
-        res = model.predict(t_eval=t_eval, inputs=inputs)
-        assert np.isinf(res).any()
+            res = model.predict(t_eval=t_eval, inputs=inputs)
+            assert np.isinf(res).any()
 
     @pytest.mark.unit
-    def test_build(self):
-        model = pybop.lithium_ion.SPM()
+    def test_build(self, model):
         model.build()
         assert model.built_model is not None
 
@@ -78,8 +84,7 @@ class TestModels:
         assert model.built_model is not None
 
     @pytest.mark.unit
-    def test_rebuild(self):
-        model = pybop.lithium_ion.SPM()
+    def test_rebuild(self, model):
         model.build()
         initial_built_model = model._built_model
         assert model._built_model is not None
@@ -200,6 +205,9 @@ class TestModels:
         expected = y0 * np.exp(-k * t_eval).reshape(-1, 1)
         solved = model.simulate(inputs, t_eval)
         np.testing.assert_array_almost_equal(solved, expected, decimal=5)
+
+        with pytest.raises(ValueError):
+            ExponentialDecay(n_states=-1)
 
     @pytest.mark.unit
     def test_basemodel(self):